Side_Projects/Shopee_Crawler_Website/target-repo/Side_Projects/Voca_Recall/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
cryptography==41.0.7
notion-client==2.2.1
python-dotenv==1.0.0
email-validator==2.0.0
APScheduler==3.10.4
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4 